+++ libs/openapi-types/src/core/mod.rs
@@ -1,0 +1,2 @@
+mod output;
+pub use output::{Output, OutputMergeOps, OutputOptionOps, OutputPairOps};
