+++ libs/openapi-types/src/lib.rs
@@ -1,5 +1,6 @@
+pub mod core;
 pub mod v3_0;
 pub mod yaml;
 
 mod error;
-pub use error::{with_key, Error, OptionOutputOps, Output, Result};
+pub use error::{with_key, Error, Output, Result};
